Add files via upload
In a bitmap index, each distinct value of a column has an associated bitmap, a string of bits where each bit corresponds to a row in the table. If the bit is set to 1, the row contains the value; if it's 0, it does not. This allows the database to quickly perform operations by checking the bitmaps instead of scanning rows.

For example, consider a table with a column that indicates whether an employee is new to the company, with values 'Yes' or 'No'. A bitmap index for this column would have two bitmaps, one for 'Yes' and one for 'No'. Each bitmap would have as many bits as there are rows in the table